Start refactoring new RegularTurn and RegularRound classes
